backend/requirements.txt
python-dotenv==0.21.1
autopep8==2.3.2
flask==3.0.3
tensorflow-cpu==2.16.1  # Use CPU version to avoid GPU issues
numpy==1.26.4
werkzeug==3.0.3
gunicorn==22.0.0
pillow==10.4.0
huggingface_hub==0.26.1
